pretty much final version
This is a local check for Check_MK written in Visual Basic. It can be used to monitor the days that have passed since the last time Sophos Endpoint Security was updated. To do this the script just checks a certain registry key in Windows. In order to check the correct key, the script checks the CPU architecture first.